% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_functions.R
\name{classify_bmi}
\alias{classify_bmi}
\title{Classify BMI into WHO Categories}
\usage{
classify_bmi(bmi)
}
\arguments{
\item{bmi}{A numeric vector of BMI scores.}
}
\value{
A factor with levels: Underweight, Normal, Overweight, Obese.
}
\description{
Classify BMI into WHO Categories
}
